refactor: изменена структура проекта части Compose экрана
